refactor: moving destructive tests to destructive folder
