+++ tests/foreman/api/test_discoveredhost.py
@@ -647,91 +647,3 @@ class TestLibvirtHostDiscovery:
                             raise AssertionError(
                                 f'Timed out waiting for {pattern[1]} from ' f'{pxe_host[0].mac}'
                             )
-
-    @pytest.mark.destructive
-    @pytest.mark.skip_if_not_set('vlan_networking')
-    def test_positive_provision_pxe_host_dhcp_change(
-        self, discovery_settings, provisioning_env, target_sat
-    ):
-        """Discovered host is provisioned in dhcp range defined in subnet entity
-
-        :id: 7ab654de-16dd-4a8b-946d-f6adde310340
-
-        :bz: 1367549
-
-        :customerscenario: true
-
-        :Setup: Provisioning should be configured and a host should be
-            discovered
-
-        :Steps:
-            1. Set some dhcp range in dhcpd.conf in satellite.
-            2. Create subnet entity in satellite with a range different from whats defined
-                in `dhcpd.conf`.
-            3. Create Hostgroup with the step 2 subnet.
-            4. Discover a new host in satellite.
-            5. Provision a host with the hostgroup created in step 3.
-
-        :expectedresults:
-            1. The discovered host should be discovered with range defined in dhcpd.conf
-            2. But provisoning the discovered host should acquire an IP from dhcp range
-                defined in subnet entity.
-
-        :CaseImportance: Critical
-        """
-        subnet = target_sat.api.Subnet(id=provisioning_env['subnet']['id']).read()
-        # Updating satellite subnet component and dhcp conf ranges
-        # Storing now for restoring later
-        old_sub_from = subnet.from_
-        old_sub_to = subnet.to
-        old_sub_to_4o = old_sub_to.split('.')[-1]
-        # Calculating Subnet's new `from` range in Satellite Subnet Component
-        new_subnet_from = subnet.from_[: subnet.from_.rfind('.') + 1] + str(int(old_sub_to_4o) - 9)
-        # Same time, calculating dhcp confs new `to` range
-        new_dhcp_conf_to = subnet.to[: subnet.to.rfind('.') + 1] + str(int(old_sub_to_4o) - 10)
-
-        cfg = get_nailgun_config()
-        with target_sat.session.shell() as shell:
-            shell.send('foreman-tail')
-            try:
-                # updating the ranges in component and in dhcp.conf
-                subnet.from_ = new_subnet_from
-                subnet.update(['from_'])
-                target_sat.execute(
-                    f'cp /etc/dhcp/dhcpd.conf /etc/dhcp/dhcpd_backup.conf && '
-                    f'sed -ie \'s/{subnet.to}/{new_dhcp_conf_to}/\' /etc/dhcp/dhcpd.conf && '
-                    f'systemctl restart dhcpd'
-                )
-                with LibvirtGuest() as pxe_host:
-                    discovered_host = _assert_discovered_host(pxe_host, shell, cfg)
-                    # Assert Discovered host discovered within dhcp.conf range before provisioning
-                    assert int(discovered_host.ip.split('.')[-1]) <= int(
-                        new_dhcp_conf_to.split('.')[-1]
-                    )
-                    # Provision just discovered host
-                    discovered_host.hostgroup = target_sat.api.HostGroup(
-                        id=provisioning_env['hostgroup']['id']
-                    ).read()
-                    discovered_host.root_pass = gen_string('alphanumeric')
-                    discovered_host.update(['hostgroup', 'root_pass'])
-                    # Assertions
-                    provisioned_host = target_sat.api.Host().search(
-                        query={
-                            'search': 'name={}.{}'.format(
-                                discovered_host.name, provisioning_env['domain']['name']
-                            )
-                        }
-                    )[0]
-                    assert int(provisioned_host.ip.split('.')[-1]) >= int(
-                        new_subnet_from.split('.')[-1]
-                    )
-                    assert int(provisioned_host.ip.split('.')[-1]) <= int(old_sub_to_4o)
-                    assert not target_sat.api.DiscoveredHost().search(
-                        query={'search': f'name={discovered_host.name}'}
-                    )
-            finally:
-                subnet.from_ = old_sub_from
-                subnet.update(['from_'])
-                target_sat.execute(
-                    'mv /etc/dhcp/dhcpd_backup.conf /etc/dhcp/dhcpd.conf /etc/dhcp/dhcpd.conf'
-                )
