+++ tests/foreman/destructive/test_foreman_rake.py
@@ -14,9 +14,9 @@
 """
 import pytest
 
+pytestmark = pytest.mark.destructive
 
-@pytest.mark.destructive
-@pytest.mark.run_in_one_thread
+
 @pytest.mark.tier3
 def test_positive_katello_reimport(target_sat):
     """Close loop bug for running katello:reimport.  Making sure
