+++ tests/foreman/destructive/test_hooks.py
